fix: Coverity CID 1675248 - Prevent unsigned integer underflow in loops
When disconnectedCount or connectedCount is 0, the expression `count - 1`
causes unsigned integer underflow because size_t is unsigned. This results
in a very large value that when assigned to int j causes undefined behavior.

Fixed by casting size_t to int before subtraction:
- `static_cast<int>(count) - 1` produces -1 when count is 0
- The loop condition `j >= 0` then fails immediately, correctly skipping
  the loop body when there's nothing to iterate

Also fixed incidental build errors:
- Added missing <iostream> include to addrman.cpp and websocket.cpp
- Fixed LogPrintf(ALL, WARNING, ...) to use WARN (matching LogLevel enum)
- Simplified setsockopt error handling to use (void) cast

+++ src/consensus/chain.cpp
@@ -343,7 +343,7 @@ bool CChainState::ActivateBestChain(CBlockIndex* pindexNew, const CBlock& block,
 
             // ROLLBACK: Reconnect all blocks we already disconnected
             std::cerr << "[Chain] ROLLBACK: Reconnecting " << disconnectedCount << " blocks..." << std::endl;
-            for (int j = disconnectedCount - 1; j >= 0; --j) {
+            for (int j = static_cast<int>(disconnectedCount) - 1; j >= 0; --j) {
                 CBlockIndex* pindexReconnect = disconnectBlocks[j];
                 CBlock reconnectBlock;
 
@@ -408,7 +408,7 @@ bool CChainState::ActivateBestChain(CBlockIndex* pindexNew, const CBlock& block,
 
             // ROLLBACK: Disconnect what we connected, reconnect what we disconnected
             std::cerr << "[Chain] ROLLBACK: Disconnecting " << connectedCount << " newly connected blocks..." << std::endl;
-            for (int j = connectedCount - 1; j >= 0; --j) {
+            for (int j = static_cast<int>(connectedCount) - 1; j >= 0; --j) {
                 if (!DisconnectTip(connectBlocks[j])) {
                     std::cerr << "[Chain] CRITICAL: Rollback failed during disconnect! Chain state corrupted!" << std::endl;
                     return false;
@@ -416,7 +416,7 @@ bool CChainState::ActivateBestChain(CBlockIndex* pindexNew, const CBlock& block,
             }
 
             std::cerr << "[Chain] ROLLBACK: Reconnecting " << disconnectedCount << " old blocks..." << std::endl;
-            for (int j = disconnectedCount - 1; j >= 0; --j) {
+            for (int j = static_cast<int>(disconnectedCount) - 1; j >= 0; --j) {
                 CBlock reconnectBlock;
 
                 // CRITICAL-C002 FIX: Explicit error handling
@@ -452,7 +452,7 @@ bool CChainState::ActivateBestChain(CBlockIndex* pindexNew, const CBlock& block,
 
             // ROLLBACK: Same as above
             std::cerr << "[Chain] ROLLBACK: Disconnecting " << connectedCount << " newly connected blocks..." << std::endl;
-            for (int j = connectedCount - 1; j >= 0; --j) {
+            for (int j = static_cast<int>(connectedCount) - 1; j >= 0; --j) {
                 if (!DisconnectTip(connectBlocks[j])) {
                     std::cerr << "[Chain] CRITICAL: Rollback failed during disconnect! Chain state corrupted!" << std::endl;
                     return false;
@@ -460,7 +460,7 @@ bool CChainState::ActivateBestChain(CBlockIndex* pindexNew, const CBlock& block,
             }
 
             std::cerr << "[Chain] ROLLBACK: Reconnecting " << disconnectedCount << " old blocks..." << std::endl;
-            for (int j = disconnectedCount - 1; j >= 0; --j) {
+            for (int j = static_cast<int>(disconnectedCount) - 1; j >= 0; --j) {
                 CBlock reconnectBlock;
 
                 // CRITICAL-C002 FIX: Explicit error handling
